[FwUpdateCfu] Update config file to improve CfuVirtualHid compatibility
The FwUpdateCfu config file is currently unable to detect & query version for the virtual HID device found in this repo (CfuVirtualHid). This PR seeks to fix this, so that the tool can at least query FW version. The PR require the PID/VID filtering fix in #64 to be merged first.

Impact to functionality:
* "FwUpdateCfu.exe version protocolCfgExample.cfg" is now able to successfully query FW version.
* "FwUpdateCfu.exe update protocolCfgExample.cfg CfuVirtualHidDevice_MCU.offer.bin CfuVirtualHidDevice_MCU.payload.bin" still fails (not a regression in this PR). This will need to be addressed later.

Changes to match defines in the CfuVirtualHid driver:
* Change PID to match PRODUCT_ID.
* Change USAGEPAGE to match CFU_DEVICE_USAGE_PAGE.
* Change USAGECOLLECTION to match CFU_DEVICE_USAGE.
* Change VERSION_FEATURE_USAGE to match VERSIONS_FEATURE_USAGE.